% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roses.R
\name{roses}
\alias{roses}
\alias{prompt}
\title{Make a roses are red poem}
\usage{
roses(
  pkg,
  hint = "",
  emoji = TRUE,
  chat = chat_openai(model = "gpt-3.5-turbo"),
  error_call = current_env()
)

prompt(pkg, hint = "", emoji = TRUE)
}
\arguments{
\item{pkg}{A package}

\item{hint}{extra information to add to the prompt}

\item{emoji}{Should the poem include emojis ?}

\item{chat}{A \link[ellmer:Chat]{ellmer::Chat} object, e.g. \code{\link[ellmer:chat_openai]{ellmer::chat_openai()}}. The default uses
the 'gpt-3.5-turbo' model from OpenAI.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A lovely poem for your package crush
}
\description{
Make a "roses are red ..." poem
about an R package.
}
\examples{
prompt("dplyr")

\dontrun{
  # this needs the OPENAI_API_KEY environment variable
  # to be set. Visit https://openai.com/api/
  roses("dplyr")
}

}
