% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roses.R
\name{roses}
\alias{roses}
\alias{prompt}
\title{Make a roses are red poem}
\usage{
roses(pkg, hint = "", emoji = TRUE, ...)

prompt(pkg, hint = "", emoji = TRUE)
}
\arguments{
\item{pkg}{A package}

\item{hint}{extra information to add to the prompt}

\item{emoji}{Should the poem include emojis ?}

\item{...}{Passed to \code{\link[openai:create_chat_completion]{openai::create_chat_completion()}}}
}
\value{
A poem generated by ChatGPT via \code{\link[openai:create_chat_completion]{openai::create_chat_completion()}}
}
\description{
Make a "roses are red ..." poem
about an R package.
}
\examples{
prompt("dplyr")

\dontrun{
  # this needs the OPENAI_API_KEY environment variable
  # to be set. See https://irudnyts.github.io/openai/
  roses("dplyr")
}
}
